42-yolo_enable_agent/requirements.txt
# YOLO Plastic Detection Enable Agent Dependencies

# Core
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.0.0

# YOLO
ultralytics>=8.0.0

# OpenAI
openai>=1.12.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization (optional)
matplotlib>=3.7.0
